requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.1
django-dotenv==1.3.0
django-filter==0.8
django-rest-framework-mongoengine==2.0.2
django-toolbelt==0.0.1
djangorestframework==3.0.4
drf-extensions==0.2.7
gunicorn==19.2.0
Markdown==2.5.1
mongoengine==0.8.7
psycopg2==2.5.4
pymongo==2.7.2
static3==0.5.1
whitenoise==1.0.6
